For corebased tests, make URLs absolute to avoid breaking when run under proxy injection mode.
The problem is that once the selenium frame infrastructure is removed (as it is under proxy injection mode), the current base URL is whatever the test page URL happens to be.  This means that relative URLs are working from the test page URL instead of the selenium frame set URL.  For the corebased tests, we are all nearly all the URLs were relative ones pointing at ../tests/html/*.html, this leads to the second open resolving to /selenium-server/tests/tests/html/..., and therefore failing.

For now I just want these tests to succeed both under and without proxy injection mode.  The larger question of what the correct behavior should be we can resolve later.


git-svn-id: http://selenium.googlecode.com/svn@2547 07704840-8298-11de-bf8c-fd130f914ac9