fix for problem where running the build from the root module would result in the distrib output not including all the non-java client drivers.
the root problem was that selenium-server was tagged to the 2.1 version of the assembly plugin and the distrib module was using 2.2-beta 1.

telling selenium-server to use the newer version however did not work for two reasons:

1) the jar-of-jars trick we use there doesn't work in 2.2-beta-1 (but does in 2.2-beta-2)
2) due to a change in the way the jar is packaged, the bouncy castle RSA signature entries were causing the resulting jar to be invalid according to java

Fixing #1 meant upgrading both distrib and selenium-server to 2.2-beta-2.

Fixing #2 was more complicated and required basically repackaging the jar near the end of the maven lifecycle (verify, which is just before install) to not include the BCKEY files that alert java to thinking there is some security signature in the MANIFEST it should be looking for


git-svn-id: http://selenium.googlecode.com/svn@4067 07704840-8298-11de-bf8c-fd130f914ac9