Fixing dogfood tests on Safari 2 by fixing isPollingForLoad.  Apparently on Safari 2, it's a TypeError to add a load listener more than once to a single iframe.  We should never have been doing this anyway, but a bug in isPollingForLoad made Safari think that we hadn't attached the listener.
The bug in isPollingForLoad: Getting properties on the frameElement would reveal properties on the frame's contentWindow, confusing isPollingForLoad's marker system.  Now we use a specially identified "frame" marker when checking to see if we've attached a load listener to the frame, so we won't get it confused with the marker we already added to the contentWindow.


git-svn-id: http://selenium.googlecode.com/svn@4275 07704840-8298-11de-bf8c-fd130f914ac9