Submits patch attached to http://jira.openqa.org/browse/SRC-665 Patch created by Chandra Sekhar Pydi
When a browser fails to launch, the exact reason for the failure is not propagated back to the user (testcase).

This patch starts to address this problem by having the Firefox browser launchers throw an exception in the case where the firefox executable could not be found, which is then in turn caught within doCommand() and the [more detailed] error message is propgated to the user instead of the previously used generic error message.



git-svn-id: http://selenium.googlecode.com/svn@6731 07704840-8298-11de-bf8c-fd130f914ac9