JasonLeyba: Updating branch (r8381:r8473).
I've tested the java, ruby, and python bindings (I don't have access to a .NET environment ATM). The chrome and remote python tests need to be updated, but that's an issue separate to this merge (they don't work with the new virtualenv setup from trunk). I will fix these in a future patch.



git-svn-id: http://selenium.googlecode.com/svn@8488 07704840-8298-11de-bf8c-fd130f914ac9