KristianRosenvold: Made portprober use platform-specific detection of ephemeral socket ports
Due to the floppy-disk themorem (the two free floppies at the back of the box are by definition defect,
otherwise they would've been in-use and hence not at the back of the box), the likelihood of getting
race conditions with ephemeral ports increases significantly the longer the
build runs. This typically results is tests failing if you run the entire suite, but not one by
one.

The current solution has a real good linux implementation, and some half-nice defaults for windows xp.

UnitTest for linux added.

git-svn-id: http://selenium.googlecode.com/svn@13813 07704840-8298-11de-bf8c-fd130f914ac9