KristianRosenvold: Fixed test failure of testShouldGetMeaningfulExceptionOnBrowserDeath
Killing firefox 3.5.19 does not work for 2.12.0 or 2.13.0, seemingly because the firefox launcher spawns another process

This test failure revealed a problem in the destroy logic on the new CommandLine implementation, which is fixed in this commit.
This error is mostly a problem of the tests and is not really expected to happen in 'live' code. Famous last words.

git-svn-id: http://selenium.googlecode.com/svn@14888 07704840-8298-11de-bf8c-fd130f914ac9