JasonLeyba: Drop JSONP for using WebDriverJs in the browser and rely solely on CORS. This means using WebDriverJs in the browser will only be supported by WebKit, Firefox, IE8+, and Opera 12+.
Unfortunately, the various browser implementations of CORS are not consistent:
* IE's XDomainRequest object may only generate requests that could be generated using a HTML form. This means it cannot generate the DELETE requests required by the JsonWireProtocol.
* WebKit's implementation of CORS does not follow the spec and forbids redirects.
* If the server returns an error response (e.g., 4xx or 5xx), IE and Opera's implementations will fire the on error handler, but without the server's response text. This renders IE and Opera incapable of handling command failures in the standard JSON protocol.

To work around this issues, I've defined a single end-point on the server to handle all CORS-based requests: /xdrpc (cross-domain rpc).  This end-point accepts POST requests with a JSON body of the form  {method:string, path:string, data:!Object}
The server will decode the command in the request body and route it to the correct command handler. The command response will always be returned as a standard JSON response object: {status:number, value:*}.  The response will always be sent with a 200 response, so clients must rely on the response's "status" field to determine whether the command succeeded.

This has been tested against Chrome, Firefox, IE, and Opera Next (12.0a)




git-svn-id: http://selenium.googlecode.com/svn@15330 07704840-8298-11de-bf8c-fd130f914ac9